Remove CNAME
Sorry guys, but this is conflicting with the docs repo, see: 
https://github.com/docker/docker.github.io/issues/14

To repro, go to http://docker.github.io/compose

This will soon be how the real URL is served at https://docs.docker.com/compose

As this is deprecated and there are a ton of links to https://docs.docker.com/compose in the docs and out in the world, it seems to make sense to alter how Fig is doing their site. 

After deleting this CNAME, I can help you guys move to another server if you need.
Another option is, if you rename this repo, the conflict would no longer exist.